+++ src/zgeqrf_mgpu.cpp
@@ -190,7 +190,7 @@ magma_zgeqrf2_mgpu(
             lapackf77_zgeqrf( &rows, &ib, hpanel(i), &ldhpanel, tau+i,
                               hwork, &lhwork, info );
             if ( *info != 0 ) {
-                fprintf( stderr, "error %d\n", (int) *info );
+                fprintf( stderr, "error %ld\n", long(*info) );
             }
 
             // Form the triangular factor of the block reflector
@@ -307,7 +307,7 @@ magma_zgeqrf2_mgpu(
         lhwork = lwork - ib*rows;
         lapackf77_zgeqrf( &rows, &ib, hwork, &rows, tau+i, hwork + ib*rows, &lhwork, info );
         if ( *info != 0 ) {
-            fprintf( stderr, "error %d\n", (int) *info );
+            fprintf( stderr, "error %ld\n", long(*info) );
         }
         
         for( j=i; j < n; j += nb ) {
