+++ src/zgesv_nopiv_batched.cpp
@@ -125,7 +125,7 @@ magma_zgesv_nopiv_batched(
     for (magma_int_t i=0; i < batchCount; i++)
     {
         if (cpu_info[i] != 0 ) {
-            printf("magma_zgetrf_batched matrix %d returned error %d\n",i, (int)cpu_info[i] );
+            printf("magma_zgetrf_batched matrix %ld returned error %ld\n",i, long(cpu_info[i]) );
             info = cpu_info[i];
             magma_free_cpu (cpu_info);
             return info;
